+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `prepare` target added inside `Makefile`
+- `GoogleDot-Red` variant added ( fixed #21 )
 
 ### Changed